renamed test data to mocks
