+++ src/__tests__/lib/containers/QueryWrapper.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallow } from 'enzyme'
 import QueryWrapper from '../../../lib/containers/QueryWrapper'
-import { mockData, mockRequest } from '../../../JSON_test_data'
+import { mockData, mockRequest } from '../../../mocks'
 
 describe('basic functionality', () => {
   let SynapseClient
