+++ src/demo/containers/Demo.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react'
 import { Link } from 'react-router-dom'
 import logoSvg from '../../images/logo.svg'
-import syn16857542Json from '../../JSON_test_data/syn16857542.json'
-import syn16858699Json from '../../JSON_test_data/syn16858699.json'
-import syn16859448Json from '../../JSON_test_data/syn16859448.json'
-import syn16859580Json from '../../JSON_test_data/syn16859580.json'
+import syn16857542Json from '../../mocks/syn16857542.json'
+import syn16858699Json from '../../mocks/syn16858699.json'
+import syn16859448Json from '../../mocks/syn16859448.json'
+import syn16859580Json from '../../mocks/syn16859580.json'
 import CustomMarkdownView from '../../lib/containers/demo_components/CustomMarkdownView'
 import Login from '../../lib/containers/demo_components/Login'
 import UserFavorites from '../../lib/containers/demo_components/UserFavorites'
